Cleaned up Env.
Most importantly, I separated the properties that used by both the boot
loader (gem5) and M3 from the ones that are only used by M3. This way,
we don't need to update gem5 whenever we change some M3-only property.

+++ src/apps/pemux/src/upcalls.rs
@@ -43,7 +43,7 @@ fn reply_msg<T>(msg: &'static tcu::Message, reply: &T) {
 fn vpe_ctrl(msg: &'static tcu::Message, state: &mut arch::State) -> Result<(), Error> {
     let req = msg.get_data::<kif::pemux::VPECtrl>();
 
-    let pe_id = req.pe_id as u32;
+    let pe_id = req.pe_id;
     let vpe_id = req.vpe_sel;
     let op = kif::pemux::VPEOp::from(req.vpe_op);
     let eps_start = req.eps_start as tcu::EpId;
