+++ src/libs/rust/base/src/io/mod.rs
@@ -70,17 +70,17 @@ macro_rules! llog {
 }
 
 #[no_mangle]
-pub extern "C" fn init_rust_io(pe_id: u32, name: *const i8) {
+pub extern "C" fn init_rust_io(pe_id: u64, name: *const i8) {
     init(pe_id, unsafe { util::cstr_to_str(name) });
 }
 
 /// Initializes the I/O module
-pub fn init(pe_id: u32, name: &str) {
+pub fn init(pe_id: u64, name: &str) {
     arch::serial::init();
     log::init(pe_id, name);
 }
 
 /// Reinitializes the I/O module (for VPE::run)
-pub fn reinit(pe_id: u32, name: &str) {
+pub fn reinit(pe_id: u64, name: &str) {
     log::reinit(pe_id, name);
 }
